Add standalone registrysyncer pkg
- And workflow handler implementation for keystone-specific logic
