+++ .changeset/young-parrots-peel.md
@@ -1,0 +1,5 @@
+---
+"chainlink": minor
+---
+
+Add registry syncer package #internal
